Rename demo-misc-vp9-opus-sw to demo-ext
