+++ demo_ext/README.md
@@ -1,0 +1,5 @@
+# Extensions Demo #
+
+A demo app that shows how to use the ExoPlayer extensions in your app.
+
+The demo app depends on the VP9, Flac and Opus Extensions being configured built correctly.
