+++ demo_ext/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.google.android.exoplayer.demo.vp9opus"
+    package="com.google.android.exoplayer.demo.ext"
     android:versionCode="1506"
     android:versionName="1.5.6"
     android:theme="@style/RootTheme">
@@ -35,7 +35,7 @@
       android:allowBackup="false"
       android:icon="@drawable/ic_launcher">
 
-    <activity android:name="com.google.android.exoplayer.demo.vp9opus.SampleChooserActivity"
+    <activity android:name="com.google.android.exoplayer.demo.ext.SampleChooserActivity"
         android:label="@string/app_name"
         android:configChanges="keyboardHidden">
       <intent-filter>
@@ -44,7 +44,7 @@
       </intent-filter>
     </activity>
 
-    <activity android:name=".PlayerActivity"
+    <activity android:name="com.google.android.exoplayer.demo.ext.PlayerActivity"
         android:configChanges="keyboardHidden|orientation|screenSize"
         android:label="@string/app_name"
         android:theme="@style/PlayerTheme"/>
