+++ demo_ext/src/main/project.properties
