+++ demo_ext/src/main/res/drawable-hdpi/ic_launcher.png
