+++ demo_ext/src/main/res/drawable-xhdpi/ic_launcher.png
