+++ demo_ext/src/main/res/drawable-xxhdpi/ic_launcher.png
