+++ demo_ext/src/main/res/drawable-xxxhdpi/ic_launcher.png
