+++ demo_ext/src/main/res/layout/activity_video_player.xml
