+++ demo_ext/src/main/res/layout/sample_chooser_inline_header.xml
