+++ demo_ext/src/main/res/values/styles.xml
