--- demo_misc/README.md
@@ -1,7 +1,0 @@
-# Miscellaneous demos #
-
-This folder contains miscellaneous demo applications. For example applications
-that demonstrate use of optional extensions, or more advanced features.
-
-A general purpose ExoPlayer demo application can be found in the [demo](../demo)
-folder.
