--- demo_misc/vp9_opus_sw/README.md
@@ -1,5 +1,0 @@
-# WebM (VP9/Opus) Software Decoder Demo #
-
-A demo app that shows how to use the ExoPlayer [VP9](../../extensions/vp9) and [Opus](../../extensions/opus) Extensions to enable VP9 and Opus playback in your app by bundling native libraries along with it.
-
-The demo app depends on the VP9 and Opus Extensions being configured built correctly.
